feat(HomeView/photo): refactoring Home view to use my photo as a separated component
